Fix possible NPE on MongoDriver#connect()
The properties can contain non-string values. If connect() is called
with such a property (e.g. intended for a different JDBC driver) the
attempt to lower-case the keys in the MongoDriver results in an NPE,
because getProperty() returns null for non-string properties.

The simplest patch is to use get() rather than getProperty().
Alternatively, to avoid the lower-casing if the URL is not actually
intended for the MongoDriver, we could add a guard clause that uses
acceptsURL().