Guard MongoStatement#setFetchSize against invalid arguments.
If a negative value is passed to setFetchSize, it leads to an internal
error down the road, which masks the original issue. To avoid this and
give immediate feedback to the client, throw an IllegalArgumentError
instead.

This occurred when migrating JDBC code from the Mongo BI Connector
(which used the MySQL JDBC driver) to MongoDB Atlas - the MySql JDBC
connector interprets Integer.MIN_VALUE as a "magic number" for setFetchSize to enable
streaming response rows.